[완료] 컴포넌트 분리 및 비교페이지 구현 - pinia를 사용한 store 데이터 저장
