+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@fortawesome/vue-fontawesome": "^3.0.3",
     "axios": "^1.3.5",
     "core-js": "^3.8.3",
+    "pinia": "^2.0.34",
     "vue": "^3.2.13",
     "vue-router": "^4.1.6"
   },
