+++ frontend/src/components/global/index.js
@@ -1,0 +1,2 @@
+
+export const URL = 'http://192.168.0.122:9000'
