+++ frontend/src/main.js
@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from '@/router/index'
+import { createPinia } from 'pinia'
 
 // awsome icon import
 import { library } from '@fortawesome/fontawesome-svg-core'
@@ -10,7 +11,10 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 
 library.add(fas, fab)
 
+const pinia = createPinia()
+
 createApp(App)
   .use(router)
+  .use(pinia)
   .component('font-awesome-icon', FontAwesomeIcon)
   .mount('#app')
