+++ src/main/java/com/imsi/car/domain/car/model/Brand.java
@@ -4,6 +4,7 @@ package com.imsi.car.domain.car.model;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
